% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{draw_lottie}
\alias{draw_lottie}
\title{Render A Lottie Animation File.}
\usage{
draw_lottie(filename)
}
\arguments{
\item{filename}{The path of a lottie file.}
}
\description{
Render A Lottie Animation File.
}
