% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vellogd-package.R
\docType{package}
\name{vellogd-package}
\alias{vellogd-package}
\title{A GPU-powered Interactive Graphics Device for R}
\description{
A blazingly fast graphics device powered by 'vello', a GPU compute-centric 2D renderer.
}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gmail.com} (\href{https://orcid.org/0000-0002-3385-7233}{ORCID})

Other contributors:
\itemize{
  \item The authors of the dependency Rust crates [contributor]
}

}
\keyword{internal}
