% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{vellogd}
\alias{vellogd}
\title{Open A 'Vello' Graphics Device.}
\usage{
vellogd(filename = "Rplot\%03d.png", width = 480, height = 480)

vellogd(filename = "Rplot\%03d.png", width = 480, height = 480)
}
\description{
Open A 'Vello' Graphics Device.

Open A 'Vello' Graphics Device With Server.
}
